+++ Sources/KeyboardKit/_Deprecated/KeyboardAction+Deprecated.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public extension KeyboardAction {
+    
+    @available(*, deprecated, message: "No keyboard actions have a standard double-tap action anymore.")
+    var standardDoubleTapAction: GestureAction? { nil }
+}